Append single char as byte (where its known to fit) to avoid (small) allocation of String on each JSON object

+++ runtime/common/src/main/kotlin/kotlinx/serialization/json/JSON.kt
@@ -186,7 +186,7 @@ data class JSON(
         fun print(v: Boolean) = engine.print(v)
 
         fun printQuoted(value: String): Unit = with(engine) {
-            print(STRING)
+            print(STRING.toByte())
             var lastPos = 0
             val length = value.length
             for (i in 0 until length) {
@@ -200,7 +200,7 @@ data class JSON(
                 lastPos = i + 1
             }
             append(value, lastPos, length)
-            print(STRING)
+            print(STRING.toByte())
         }
     }
